refs #166 @2h
There are more then one issue here. The test cases were passing the address of the pthread::runnable instance. When this reference was passed to the pthread::thread constructor, it was dereferenced it agin ending somewhere in unallocated memory. So the first thing was to use std::unique_ptr and this seems to fix things.

When creating a pthread::thread, the pthread attributes are not created. When the destructor of such an instance is called it tries to destroy this never alocated attribute. The program core dumps...